backend_alt/requirements.txt
bidict==0.23.1
click==8.3.2
dnspython==2.8.0
greenlet==3.4.0
h11==0.16.0
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
uvicorn==0.44.0
wsproto==1.3.2
